+++ modern-data-platform-stack/static-data/scripts/platys-mdp-docker-image-deploy/docker-entrypoint.sh
@@ -1,0 +1,35 @@
+#!/bin/bash
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+set -e
+
+if [[ "$VERBOSE" == "yes" ]]; then
+    set -x
+fi
+
+# if the DOCKER_REGISTRY is set
+if [[ -v DOCKER_REGISTRY && -n "$DOCKER_REGISTRY" ]]; then
+    apk add --no-cache wait4x 
+
+    # wait until docker registry is avalable
+    wait4x http http://$DOCKER_REGISTRY --insecure-skip-tls-verify --timeout 400s --interval 5s
+fi
+
+# run all the shell scripts in the init folder once
+source run-init
+
+# execute command passed in as arguments.
+exec "$@"