+++ documentation/changes.md
@@ -49,6 +49,7 @@ The Modern Data Platform version 1.18.0 contains the following bug fixes and enh
 
  * `minio-console` removed as docker image is no longer available
  * `MINIO_browser_enable` renamed to `MINIO_browser_enabled` to be conforment with naming convention
+ * Change internal port of `docker-registry` from `5000` to `5020`
 
 ## What's new in 1.18.1